docker/computational-complexity-linux/requirements.txt
scdiffeq==v0.1.8
scdiffeq-analyses==v0.0.2rc5
wandb
pyyaml
psutil